Pass feed URL straight through to govukdelivery.
This greatly simplifies the process of creating
email signups.

Rather than passing params to a form which sends
options a controller which then regenerates an
atom URL by parts, just pass the URL straight
through and ask the user only for a confirmation
of their subscription (with local government
opt-in).